Separation of merging algorithms
