--- Detector/DetComponents/tests/options/mergeCellsFromList.py
@@ -1,54 +1,0 @@
-from Gaudi.Configuration import *
-from Configurables import ApplicationMgr, HepMCReader, HepMCDumper
-
-reader = HepMCReader("Reader", Filename="/afs/cern.ch/exp/fcc/sw/0.7/testsamples/testHepMCrandom.dat")
-reader.DataOutputs.hepmc.Path = "hepmc"
-
-from Configurables import HepMCConverter
-hepmc_converter = HepMCConverter("Converter")
-hepmc_converter.DataInputs.hepmc.Path="hepmc"
-hepmc_converter.DataOutputs.genparticles.Path="allGenParticles"
-hepmc_converter.DataOutputs.genvertices.Path="allGenVertices"
-
-dumper = HepMCDumper("Dumper")
-dumper.DataInputs.hepmc.Path="hepmc"
-
-from Configurables import GeoSvc
-geoservice = GeoSvc("GeoSvc",  detectors=['file:Detector/DetFCChhBaseline1/compact/FCChh_DectEmptyMaster.xml',
-                                         'file:Detector/DetFCChhECalSimple/compact/FCChh_ECalBarrel_Mockup.xml' ],# detectors=['file:Test/TestGeometry/data/TestBoxCaloSD_3readouts.xml'],
-                    OutputLevel = DEBUG)
-
-from Configurables import SimG4Svc
-geantservice = SimG4Svc("SimG4Svc", physicslist='SimG4TestPhysicsList')
-
-from Configurables import SimG4Alg, SimG4SaveCalHits, InspectHitsCollectionsTool
-inspecttool = InspectHitsCollectionsTool("inspect", readoutNames=["ECalHitsPhiEta"], OutputLevel = DEBUG)
-savecaltool = SimG4SaveCalHits("saveECalHits", readoutNames = ["ECalHitsPhiEta"], OutputLevel = DEBUG)
-savecaltool.DataOutputs.caloClusters.Path = "caloClusters"
-savecaltool.DataOutputs.caloHits.Path = "caloHits"
-geantsim = SimG4Alg("SimG4Alg", outputs= ["SimG4SaveCalHits/saveECalHits","InspectHitsCollectionsTool/inspect"])
-
-from Configurables import MergeCells
-merge = MergeCells("mergeCells",
-                   # take the bitfield description from the geometry service
-                   readout ="ECalHitsPhiEta",
-                   # cells in which field should be merged
-                   identifier = "phi",
-                   # how many cells to merge
-                   # merge first 20 into new cell (id=0), next 50 into second cell (id=1), ...
-                   mergeList = [20,50,47],
-                   volumeMatchName = "LAr",
-                   OutputLevel = DEBUG)
-merge.DataInputs.inhits.Path = "caloHits"
-merge.DataOutputs.outhits.Path = "newCaloHits"
-
-from Configurables import FCCDataSvc, PodioOutput
-podiosvc = FCCDataSvc("EventDataSvc")
-out = PodioOutput("out", filename="testMergeCells.root")
-out.outputCommands = ["keep *"]
-
-ApplicationMgr(EvtSel='NONE',
-               EvtMax=30,
-               TopAlg=[reader, hepmc_converter, geantsim, merge, out],
-               ExtSvc = [podiosvc, geoservice, geantservice],
-               OutputLevel=DEBUG)
